misc cleanup / refactoring
