+++ cmd/analyzer/main.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/buildpack/lifecycle"
 	"github.com/buildpack/lifecycle/cmd"
-	"github.com/buildpack/lifecycle/docker"
 	"github.com/buildpack/lifecycle/image/auth"
 )
 
@@ -95,7 +94,7 @@ func analyzer() error {
 	var img imgutil.Image
 
 	if useDaemon {
-		dockerClient, err := docker.DefaultClient()
+		dockerClient, err := cmd.DockerClient()
 		if err != nil {
 			return cmd.FailErr(err, "create docker client")
 		}
